pyproject.toml
[project]
name = "doc-reading-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "camel-ai>=0.2.82",
    "flash-attn>=2.8.3",
    "loguru>=0.7.3",
    "pdf2image>=1.17.0",
    "pymilvus>=2.6.6",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.0",
    "qwen-vl-utils>=0.0.14",
    "torch==2.8.0",
    "transformers>=4.57.3",
    "vllm>=0.11.0",
]
